changed datatypes to double

+++ sqllab/createtable.sql
@@ -1,9 +1,9 @@
 DROP TABLE IF EXISTS earthquakes;
 CREATE TABLE earthquakes (
-  latitude double,
-  longitude double,
-  depth double,
-  magnitude double,
+  latitude double precision,
+  longitude double precision,
+  depth double precision,
+  magnitude double precision,
   magnitudeType text,
   id text,
   locationSource text
